Rename exercises base folder to "practice".
Update dependencies. Create new shortcuts @.
