+++ src/app/components/navbar/components/settings-popup/settings-popup.actions.ts
@@ -1,6 +1,7 @@
-import { ExerciseModeType, TestingAgainstType } from '@app/pages/classroom/exercises/store';
-import { ThemeType } from '@app/store/app';
 import { createActionGroup, emptyProps, props } from '@ngrx/store';
+
+import { ExerciseModeType, TestingAgainstType } from '@practice/store';
+import { ThemeType } from '@app/store/app';
 
 
 export const SettingsPopupAction = createActionGroup(
