--- src/app/pages/classroom/exercises/store/exercises/index.ts
@@ -1,5 +1,0 @@
-
-export * from './exercises.actions';
-export * from './exercises.reducer';
-export * from './exercises.effects';
-export * from './exercises.selectors';
