--- src/app/pages/classroom/exercises/store/index.ts
@@ -1,1 +1,0 @@
-export * from './exercises';
