+++ src/app/pages/classroom/practice/pages/exercises/components/modes/quiz/quiz.component.scss
