+++ src/app/pages/classroom/practice/pages/exercises/components/modes/spelling/spelling.component.scss
