+++ src/app/pages/classroom/practice/pages/exercises/components/results/results.component.scss
