+++ src/app/pages/classroom/practice/pages/exercises/shared/directives/auto-focus.directive.ts
