+++ src/app/pages/classroom/practice/pages/exercises/shared/progress-bar/progress-bar.component.ts
