+++ src/app/pages/classroom/practice/pages/exercises/utils/shuffleArray.ts
