+++ src/app/pages/classroom/practice/pages/exercises/utils/switchCase.ts
