+++ src/app/pages/classroom/practice/pages/select-group/select-group.component.scss
