+++ src/app/pages/classroom/practice/practice-menu-routing.module.ts
@@ -1,13 +1,13 @@
 import { Route } from '@angular/router';
 
 import { ExerciseContainerComponent } from './pages/exercises/exercise-container.component';
-import { ExerciseMenuComponent } from './exercise-menu.component';
+import { ExerciseMenuComponent } from './practice-menu.component';
 import { SelectGroupComponent } from './pages/select-group/select-group.component';
 import { StoreModule, provideState, provideStore } from '@ngrx/store';
-import { exercisesFeature } from './store/exercises/exercises.reducer';
+import { exercisesFeature } from './store/practice/practice.reducer';
 import { importProvidersFrom } from '@angular/core';
 import { EffectsModule } from '@ngrx/effects';
-import { ExercisesEffects } from './store/exercises/exercises.effects';
+import { ExercisesEffects } from './store/practice/practice.effects';
 
 export default [
   {
