+++ src/app/pages/classroom/practice/practice-menu.component.html
