+++ src/app/pages/classroom/practice/practice-menu.component.scss
