+++ src/app/pages/classroom/practice/store/index.ts
@@ -1,0 +1,1 @@
+export * from './practice';
