+++ src/app/pages/classroom/practice/store/practice/index.ts
@@ -1,0 +1,5 @@
+
+export * from './practice.actions';
+export * from './practice.reducer';
+export * from './practice.effects';
+export * from './practice.selectors';
