+++ src/app/pages/classroom/practice/store/practice/practice.selectors.ts
@@ -1,5 +1,5 @@
 import { createSelector } from "@ngrx/store";
-import { selectCurrentWord, selectAnswerInput, selectIsActionNext, selectIsActionProofread, selectIsLastAnswerCorrect, selectProgress, selectAnswerLocked, selectTestingAgainst, selectExerciseWords, selectAnswerChoices } from "./exercises.reducer";
+import { selectCurrentWord, selectAnswerInput, selectIsActionNext, selectIsActionProofread, selectIsLastAnswerCorrect, selectProgress, selectAnswerLocked, selectTestingAgainst, selectExerciseWords, selectAnswerChoices } from "./practice.reducer";
 import { _getAnswers, _isResponseCorrect } from "./store-utils";
