+++ tsconfig.json
@@ -8,8 +8,10 @@
       "@shared/*":["app/shared/*"],
       "@services/*":["app/shared/services/*"],
       "@env/*":["environments/*"],
+      "@classroom/*":["app/pages/classroom/*"],
       "@glossary/*":["app/pages/classroom/glossary/*"],
-      "@exercises/*":["app/pages/classroom/exercises/*"],
+      "@practice/*":["app/pages/classroom/practice/*"],
+      "@exercises/*":["app/pages/classroom/practice/pages/exercises/*"],
     },
     "outDir": "./dist/out-tsc",
     "sourceMap": true,
